Run As Context Menu shows Ant Run Configurations multiple times #527
AntLaunchShortcutWithDialog now implements only ILaunchShortcut
interface.
The common implementation of AntLaunchShortcut and
AntLaunchShortcutWithDialog are moved to AntLaunchShortcutCommon class

Fixes https://github.com/eclipse-platform/eclipse.platform/issues/527
